% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.tsdata}
\alias{print.tsdata}
\title{Print method for tsdata objects}
\usage{
\method{print}{tsdata}(x, ...)
}
\arguments{
\item{x}{A tsdata object created by \code{tsdata()}.}

\item{...}{Additional arguments.}
}
\description{
This method prints a compact summary of a \code{tsdata} object.
}
