% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.tsdata}
\alias{summary.tsdata}
\title{Summary method for tsdata objects}
\usage{
\method{summary}{tsdata}(object, ...)
}
\arguments{
\item{object}{A tsdata object created by \code{tsdata()}.}

\item{...}{Additional arguments.}
}
\value{
A list with summary statistics.
}
\description{
This method returns a summary of a \code{tsdata} object.
}
