% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsdata.R
\name{tsdata}
\alias{tsdata}
\title{Create a tsdata time series object}
\usage{
tsdata(data, dates)
}
\arguments{
\item{data}{Numeric vector of values}

\item{dates}{Date vector of same length as data}
}
\value{
A tsdata object
}
\description{
Create a tsdata time series object
}
